adding fontawesome library icons in icon.ts
